skip OSImageStream reconciliation in OKD
